+++ modules/features/taskerplugin/src/main/java/au/com/shiftyjelly/pocketcasts/taskerplugin/controlplayback/ActionRunnerControlPlayback.kt
@@ -86,7 +86,7 @@ class ActionRunnerControlPlayback : TaskerPluginRunnerActionNoOutput<InputContro
         if (overrideGlobalEffects) {
             podcastManager.updateEffects(currentPodcast, playbackEffects)
         } else {
-            settings.globalPlaybackEffects.set(playbackEffects, needsSync = true)
+            settings.globalPlaybackEffects.set(playbackEffects, updateModifiedAt = true)
         }
         playbackManager.updatePlayerEffects(playbackEffects)
     }