+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/MediaSessionManager.kt
@@ -761,7 +761,7 @@ class MediaSessionManager(
             // update global playback speed
             val effects = settings.globalPlaybackEffects.value
             effects.playbackSpeed = newSpeed
-            settings.globalPlaybackEffects.set(effects, needsSync = true)
+            settings.globalPlaybackEffects.set(effects, updateModifiedAt = true)
             playbackManager.updatePlayerEffects(effects = effects)
         }
     }