+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/PodcastManagerImpl.kt
@@ -594,7 +594,7 @@ class PodcastManagerImpl @Inject constructor(
 
     override suspend fun updateAllShowNotifications(showNotifications: Boolean) {
         if (showNotifications) {
-            settings.notifyRefreshPodcast.set(true, needsSync = true)
+            settings.notifyRefreshPodcast.set(true, updateModifiedAt = true)
         }
         podcastDao.updateAllShowNotifications(showNotifications)
     }
@@ -656,7 +656,7 @@ class PodcastManagerImpl @Inject constructor(
 
     override fun updateShowNotifications(podcast: Podcast, show: Boolean) {
         if (show) {
-            settings.notifyRefreshPodcast.set(true, needsSync = true)
+            settings.notifyRefreshPodcast.set(true, updateModifiedAt = true)
         }
         podcastDao.updateShowNotifications(show, podcast.uuid)
     }